Working on cpp bindings
Former-commit-id: 7a622d8099e15534f89d75ccd0405a1683383603

+++ cpp_bindings/MLFImage.ml
@@ -58,7 +58,7 @@ let _ =
   Callback.register "makeLE" (fun a b -> Cmp (LE, a, b));
   Callback.register "makeSelect" (fun c a b -> Select (c, a, b));
   Callback.register "makeVar" (fun a -> Var a);
-  Callback.register "makeLoad" (fun buf idx -> Load (i32, {buf=buf; idx=idx}));
+  Callback.register "makeLoad" (fun buf idx -> Load (f32, {buf=buf; idx=idx}));
   Callback.register "makeStore" (fun a buf idx -> Store (a, {buf=buf; idx=idx}));
   Callback.register "makeFunction" (fun args stmt -> ((List.rev args), stmt));
   Callback.register "makeMap" (fun var min max stmt -> Map (var, min, max, stmt));
